# apd: don't strip trailing zeros in Quo and Sqrt