--- a/example_test.go
+++ b/example_test.go
@@ -56,9 +56,10 @@ func ExampleContext_inexact() {
 			return
 		}
 	}
-	// Output: d:       9, inexact: false, err: <nil>
-	// d:       3, inexact: false, err: <nil>
-	// d:       1, inexact: false, err: <nil>
+	// Output:
+	// d:  9.0000, inexact: false, err: <nil>
+	// d:  3.0000, inexact: false, err: <nil>
+	// d:  1.0000, inexact: false, err: <nil>
 	// d: 0.33333, inexact:  true, err: <nil>
 }