--- a/gda_test.go
+++ b/gda_test.go
@@ -711,10 +711,6 @@ func (tc TestCase) PrintIgnore() {
 }
 
 var GDAignore = map[string]bool{
-	// GDA says decNumber should skip these
-	"powx4302": true,
-	"powx4303": true,
-
 	// Invalid context
 	"expx901":  true,
 	"expx902":  true,
@@ -775,21 +771,6 @@ var GDAignore = map[string]bool{
 
 	// TODO(mjibson): fix tests below
 
-	// log10(x) with large exponents, overflows
-	"logx0020": true,
-	"logx1146": true,
-	"logx1147": true,
-	"logx1156": true,
-	"logx1157": true,
-	"logx1176": true,
-	"logx1177": true,
-
-	// log10(x) where x is 1.0 +/- some tiny epsilon. Our results are close but
-	// differ in the last places.
-	"logx1304": true,
-	"logx1309": true,
-	"logx1310": true,
-
 	// overflows to infinity
 	"powx4125": true,
 	"powx4145": true,
@@ -819,90 +800,13 @@ var GDAignore = map[string]bool{
 	"addx61618": true,
 
 	// extreme input range, but should work
-	"lnx0902":  true,
-	"sqtx8137": true,
-	"sqtx8139": true,
-	"sqtx8145": true,
-	"sqtx8147": true,
-	"sqtx8149": true,
-	"sqtx8150": true,
-	"sqtx8151": true,
-	"sqtx8158": true,
-	"sqtx8165": true,
-	"sqtx8168": true,
-	"sqtx8174": true,
-	"sqtx8175": true,
-	"sqtx8179": true,
-	"sqtx8182": true,
-	"sqtx8185": true,
-	"sqtx8186": true,
-	"sqtx8195": true,
-	"sqtx8196": true,
-	"sqtx8197": true,
-	"sqtx8199": true,
-	"sqtx8204": true,
-	"sqtx8212": true,
-	"sqtx8213": true,
-	"sqtx8214": true,
-	"sqtx8218": true,
-	"sqtx8219": true,
-	"sqtx8323": true,
-	"sqtx8324": true,
-	"sqtx8331": true,
-	"sqtx8626": true,
-	"sqtx8627": true,
-	"sqtx8628": true,
-	"sqtx8631": true,
-	"sqtx8632": true,
-	"sqtx8633": true,
-	"sqtx8634": true,
 	"sqtx8636": true,
-	"sqtx8637": true,
-	"sqtx8639": true,
-	"sqtx8640": true,
-	"sqtx8641": true,
 	"sqtx8644": true,
-	"sqtx8645": true,
 	"sqtx8646": true,
 	"sqtx8647": true,
 	"sqtx8648": true,
-	"sqtx8649": true,
 	"sqtx8650": true,
 	"sqtx8651": true,
-	"sqtx8652": true,
-	"sqtx8654": true,
-
-	// tricky cases of underflow subnormals
-	"sqtx8700": true,
-	"sqtx8701": true,
-	"sqtx8702": true,
-	"sqtx8703": true,
-	"sqtx8704": true,
-	"sqtx8705": true,
-	"sqtx8706": true,
-	"sqtx8707": true,
-	"sqtx8708": true,
-	"sqtx8709": true,
-	"sqtx8710": true,
-	"sqtx8711": true,
-	"sqtx8712": true,
-	"sqtx8713": true,
-	"sqtx8714": true,
-	"sqtx8715": true,
-	"sqtx8716": true,
-	"sqtx8717": true,
-	"sqtx8718": true,
-	"sqtx8719": true,
-	"sqtx8720": true,
-	"sqtx8721": true,
-	"sqtx8722": true,
-	"sqtx8723": true,
-	"sqtx8724": true,
-	"sqtx8725": true,
-	"sqtx8726": true,
-	"sqtx8727": true,
-	"sqtx8728": true,
-	"sqtx8729": true,
 }
 
 var GDAignoreFlags = map[string]bool{
@@ -915,9 +819,4 @@ var GDAignoreFlags = map[string]bool{
 	"sqtx9039": true,
 	"sqtx9040": true,
 	"sqtx9045": true,
-
-	// missing underflow, subnormal
-	"expx048": true,
-	"expx756": true,
-	"expx757": true,
 }